+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.10 - Unreleased
 
+### Added
+* Ability to specify `use_native_scale` in the tr dict, which returns jagged data at its original resolution
+* Ability to specify `scale` which augments the scale of the data.
+
 
 ## Version 0.6.9 - Released 2022-07-19